Added model class representing authMethod configuration, added license text.
