+++ powerauth-webflow-authentication/src/main/java/io/getlime/security/powerauth/lib/webflow/authentication/service/AuthMethodQueryService.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2017 Lime - HighTech Solutions s.r.o.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.getlime.security.powerauth.lib.webflow.authentication.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -6,6 +21,7 @@ import io.getlime.powerauth.soap.ActivationStatus;
 import io.getlime.powerauth.soap.GetActivationListForUserResponse;
 import io.getlime.security.powerauth.lib.nextstep.client.NextStepClient;
 import io.getlime.security.powerauth.lib.nextstep.client.NextStepServiceException;
+import io.getlime.security.powerauth.lib.nextstep.model.entity.AuthMethodConfiguration;
 import io.getlime.security.powerauth.lib.nextstep.model.entity.UserAuthMethodDetail;
 import io.getlime.security.powerauth.lib.nextstep.model.enumeration.AuthMethod;
 import io.getlime.security.powerauth.lib.nextstep.model.response.GetUserAuthMethodsResponse;
@@ -80,8 +96,8 @@ public class AuthMethodQueryService {
                 String config = authMethodDetail.getConfig();
                 if (config != null && !config.isEmpty()) {
                     try {
-                        Activation activation = objectMapper.readValue(config, Activation.class);
-                        String activationId = activation.getActivationId();
+                        AuthMethodConfiguration configuration = objectMapper.readValue(config, AuthMethodConfiguration.class);
+                        String activationId = configuration.getActivationId();
                         if (activationId != null && !activationId.isEmpty()) {
                             // set successfully parsed activationId from configuration
                             configuredActivation = activationId;
@@ -117,18 +133,4 @@ public class AuthMethodQueryService {
         return false;
     }
 
-    /**
-     * Helper class used during deserialization from JSON.
-     */
-    private class Activation {
-        private String activationId;
-
-        public String getActivationId() {
-            return activationId;
-        }
-
-        public void setActivationId(String activationId) {
-            this.activationId = activationId;
-        }
-    }
 }
